Rework exception reporting
Instead of using `error` with some `String` message, add custom
exception types and use `MonadThrow` from the `exceptions` package to
report errors.

This commit adds some useful functionality related to the `vector`
package to handle exceptions in `ST` actions used with `create`,
`modify` and similar functions. To ease working in the `CatchST` monad
introduced for this purpose, which is a monad transformer, the
`Data.Vector.Generic.Lifted` module is introduced which exports lifted
versions of mutable vector manipulation actions.