Windows: Enable SIMD/AVX
This commit changes `galois_amd64.c` to only compile the AVX version of
the routine, and let all exported implementations use this under the
hood.

Rationale: MinGHC on Windows comes with GCC 4.5.2 which doesn't provide
what's required for the other implementations, at all. Once the GCC
version shipped with Stack on Windows is bumped, the implementations as
done on Linux could be used again.