Add `SIMD-SSE` flag
On TravisCI using MacOS X AVX or SSSE3 instructions seem to be
unsupported. This flag allows to disable AVX/SSSE3 instruction
generation while compiling the routines in `galois_amd64.c`, and
fallback to SSE4 instructions.

This is not applicable to Linux systems, which bring several
target-specific implementations, dispatched at runtime.