cbits: Add AVX2 routines
This commit adds hand-rolled AVX2 implementations of the Galois-field
multiplication routines, based on the standard template.

`ifunc` resolution, benchmarks and tests are updated accordingly.

No compiler-generated version is provided (yet) because the compiler
insists on using SIMD instructions which require aligned access, whilst
the only guarantee we have is for vectors to be 8 (or 16?) byte aligned.
This results in crashes (`SIGSEGV`, as expected).