Rework `cbits` parts
This is a major rework of the `cbits` part, mainly bringing support for
all different SIMD routines to all supported platforms.

The number of supported levels of SIMD extensions is also changed
somewhat: from AVX2, AVX, SSE4.1 and generic to AVX2, AVX, SSSE3, SSE2
and generic. This is more in line with the actually used instructions.

Note AVX2 support depends on the compiler supporting this instruction
set, so it might not be available on all platforms.

Because of a major rework of the implementation of the SIMD routines,
using inlined functions instead of macros, there are no 'optimized' and
'compiler' versions any longer: everything is optimized depending on
available chipset support in a compilation target.

Because instruction availability support is now determined at runtime
for all supported platforms, the `SIMD-SSE` Cabal flag is removed.

Note, because of autotools, one should now run something along the
lines of `autoreconf -vi` inside the `cbits` directory when building
this package from VCS. Release packages should contain all required
files.

In line with the above, all CI integration scripts were updated to run
the required pre-build commands, including installation of any
dependencies.