Force `autoreconf` to regenerate all files during `sdist`
When files are not regenerated, some (unchanged) files like
`config.h.in` could be kept as-is in a dirty tree, which causes `make`
to try to generate all files again (due to 'maintainer mode'), which in
turn depends on `aclocal` and other tools.

Forcing the files to be generated during `sdist` should fix all
timestamps, and creates a correct tarball which can be built without
extra tools on the system.