Build: Use `ar-wrapper`
The easiest way to convince Cabal to add a couple of 'foreign' object
files into the resulting library archive is to wrap the `ar` command and
modify the command line on the fly.

Not very portable, not very friendly, but for now there's no better way
(unless copying half of Cabal into `Setup.hs`...).