cbits: Add tests for foreign ISA SIMD routines
This commit adds a couple of tests to validate SIMD routine
implementations targetting an ISA other than the one of the host system.

First, an executable is added (`reedsolomon-gal-mul-stdio`) which reads
the `low_vector`, `high_vector` and `in` arguments of
`reedsolomon_gal_mul` from `stdin`, and dumps the result to `stdout`.

Secondly, a small Haskell script is added which takes two commands on
the command line, which should adhere to the protocol expected by
`reedsolomon-gal-mul-stdio`. The script will use QuickCheck to generate
random inputs (for `low_vector`, `high_vector` and `in`), call both
commands and assert the result data is the same.

Note: this script has a couple of dependencies not listed in
`reedsolomon.cabal`, including `process-extras`. If these dependencies
are not yet available on the system, they can be installed through
`stack install`.

Finally, a utility script is added to build `reedsolomon-gal-mul-stdio`
for the host system and a foreign ISA (currently hard-coded to
`armv7-rpi2-linux-gnueabihf`), and then run
`reedsolomon-gal-mul-stdio-quickcheck` on both of them, one natively and
one using the Qemu user-space emulation of the foreign ISA (currently
hard-coded to using `qemu-arm`).

The approach above was taken to validate the ARM support added in
6282161457 is most likely to work correctly.

See: 6282161457cc778dded715c3e6b7e167f6557cbc