cbits: Fix detection of `-floop-parallelize-all`
On older systems and GCC versions, `-floop-parallelize-all` depends on
some runtime libraries which might not be installed (e.g. on Ubuntu /
TravisCI).

The previous check succeeded because the error is only triggered when
the optimization actually fires, which is not the case for the default
source snippet used to test compilation.

This new version should fix this.

See: https://lists.debian.org/debian-gcc/2015/06/msg00164.html