cbits: Add `rpl_malloc` implementation
Due to the use of `AC_FUNC_MALLOC` in `configure.ac` (because
`reedsolomon-gal-mul-stdio` uses `malloc`), the `configure` script
checks whether the system `malloc` is according to GNU libc: returning a
valid pointer when allocating 0 bytes, and setting `errno` to `ENOMEM`
on allocation failure.

When cross-compiling, this check can't be executed and the script
pessimistically assumes the target system's `malloc` is foreign. It then
creates a `define` from `malloc` to `rpl_malloc`, which by default
doesn't exist. The `LIBOBJS` variable is extended with `malloc.o` in
order to be able to provide a source file which contains this
function/symbol, which is exactly what this patch does.

There's no use of `malloc` and friends in the SIMD routine code, so this
should have no impact on encoding/decoding of data.