cbits: Implement PowerPC AltiVec support
This commit adds PowerPC AltiVec support to `reedsolomon.c` within the
existing framework.

Note two slightly different implementations are provided, dispatched in
a `configure` check: one using the built-in compiler support for
right-shift of a vector (which should work on any target system),
another one using the low-level `vec_vsrd` intrinsic, which is only
available on POWER8 systems. As such it is important to export the
right flags (e.g. `CPPFLAGS='-mcpu=power8'`) when building the package.

Unlike the Intel implementations, no runtime support check is performed,
and using the library on an incompatible system will result in crashes.