TravisCI: Update `qemu-user-static` package retrieval
The previously hard-coded package version was updated on the Ubuntu
mirrors, and with it the original package version has been removed,
causing the build to fail.

This commit updates the package identifier to the latest uploaded
version, and points to `launchpadlibrarian.net` instead, so this should
be retrievable in the foreseeable future.