cbits: Check for `#pragma clang loop unroll` support
Some older Clang versions, e.g. the one running on TravisCI's OSX build
machines, don't support the `clang loop unroll` pragma, which in turn
triggers a compiler error.

There seems to be no way to use `__has_feature` to detect whether the
pragma is available, so adding a test to `configure` is the way to go.