cbits: Enable binary IO in `reedsolomon-gal-mul-stdio` on Win32
By default, `stdin` and `stdout` aren't set to binary mode on Win32,
causing some characters to be interpreted/replaced (e.g. `\n` to be
converted in `\r\n`). This obviously breaks the (QuickCheck) testing of
`reedsolomon-gal-mul-stdio`.

This patch adds some Win32-specific code to set the file mode of `stdin`
and `stdout` to `_O_BINARY` accordingly.