Tests: Add vector address alignment assertion test
In the `cbits` code, we currently assume input vectors to the
`reedsolomon_galois_mul*` procedurues are 16-byte aligned (defined by
`_RS_ASSUMED_ALIGNMENT` in `configure.ac`). This should be true, given
how storable vectors are allocated, and the GHC RTS always ensuring
16-byte alignment (see
https://github.com/ghc/ghc/blob/d3a79bcc035d7dab4f16a807984225e4bd4fccd8/rts/PrimOps.cmm#L71).

To ensure this remains the case in future GHC versions, or whichever
version used with this library (though anything after somewhere 2009
should be OK), this commit adds a simple test-case which allocates a
bunch of vectors, and validates their address.