cbits: Turn some warnings into errors
In release builds, `-Werror` is not added to `CFLAGS`. This causes some
checks used to validate whether some functionality is available on the
system to generate a warning and succeed, then generating a warning
during compilation as well.

This patch adds some `-Werror=...` for warnings which we want to be
errors.