Hack `-fno-warn-redundant-constraints` into `ghc-wrapper`
Turns out adding the flag to `Ghc-Options` in the Cabal configuration
file is not sufficient when using `stack build --pedantic`: Stack adds
`-Wall -Werror` to the command-line options passed to GHC eventually,
which discards any warning-disabling-flags passed *before* said options.

The only way around seems to be to add a hack into `ghc-wrapper` to
append the option when GHC 8 is used. Hmpf.