cbits: Detect and use system-provided `__get_cpuid_count`
Starting in GCC 6.3, the `cpuid.h` header provided by the compiler
defines a `__get_cpuid_count` function. If this is the case, redefining
it in one of our modules (`reedsolomon_dispatch.c`) causes compilation
to fail.

This commit adds a check for this declaration in `cpuid.h` to
`configure.ac` so a relevant define is generated at `configure` time,
and moves our definition of `__get_cpuid_count` inside a conditional
block.